pyproject.toml
[project]
name = "datacontract-cli"
version = "0.9.7"
description = "Test data contracts"
readme = "README.md"
authors = [
  { name = "Jochen Christ", email = "jochen.christ@innoq.com" },
  { name = "Stefan Negele", email = "stefan.negele@innoq.com" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
  "typer[all]~=0.9.0",
  "pydantic>=2.5.3,<2.7.0",
  "pyyaml~=6.0.1",
  "requests~=2.31.0",
  "fastapi==0.110.0",
  "fastparquet==2024.2.0",
  "python-multipart==0.0.9",
  "rich~=13.7.0",
  "simple-ddl-parser==1.0.3",
  "soda-core-bigquery~=3.2.1",
  "soda-core-duckdb~=3.2.1",
  "soda-core-postgres~=3.2.1",
  "soda-core-snowflake~=3.2.1",
  "soda-core-spark[databricks]~=3.2.1",
  "soda-core-spark-df~=3.2.1",
  "snowflake-connector-python[pandas]>=3.6,<3.8",
  "duckdb==0.10.0",
  "fastjsonschema~=2.19.1",
  "python-dotenv~=1.0.0",
  "s3fs==2024.3.1",
  "rdflib==7.0.0",
  "avro==1.11.3",
  "opentelemetry-exporter-otlp-proto-grpc~=1.16.0",
  "opentelemetry-exporter-otlp-proto-http~=1.16.0"
]

[project.optional-dependencies]
dev = [
  "httpx==0.27.0",
  "ruff",
  "pytest",
  "testcontainers-minio",
  "testcontainers-postgres",
  "testcontainers-kafka",
]

[project.urls]
Homepage = "https://cli.datacontract.com"
Issues = "https://github.com/datacontract/cli/issues"

[project.scripts]
datacontract = "datacontract.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120